+++ JsonToolsNppPlugin/Forms/FindReplaceForm.Designer.cs
@@ -47,8 +47,8 @@
             this.AdvancedGroupBox = new System.Windows.Forms.GroupBox();
             this.MatchExactlyBox = new System.Windows.Forms.CheckBox();
             this.IgnoreCaseCheckBox = new System.Windows.Forms.CheckBox();
-            this.AdvancedGroupBoxLabel = new System.Windows.Forms.Label();
             this.SwapFindReplaceButton = new System.Windows.Forms.Button();
+            this.ShowAdvancedOptionsCheckBox = new System.Windows.Forms.CheckBox();
             this.AdvancedGroupBox.SuspendLayout();
             this.SuspendLayout();
             // 
@@ -62,9 +62,10 @@
             this.KeysValsBothBox.Location = new System.Drawing.Point(8, 21);
             this.KeysValsBothBox.Name = "KeysValsBothBox";
             this.KeysValsBothBox.Size = new System.Drawing.Size(133, 24);
-            this.KeysValsBothBox.TabIndex = 3;
-            this.KeysValsBothBox.TabStop = false;
+            this.KeysValsBothBox.TabIndex = 5;
             this.KeysValsBothBox.Visible = false;
+            this.KeysValsBothBox.KeyDown += new System.Windows.Forms.KeyEventHandler(this.FindReplaceForm_KeyDown);
+            this.KeysValsBothBox.KeyPress += new System.Windows.Forms.KeyPressEventHandler(this.TextBox_KeyPress);
             this.KeysValsBothBox.KeyUp += new System.Windows.Forms.KeyEventHandler(this.FindReplaceForm_KeyUp);
             // 
             // RecursiveSearchBox
@@ -75,8 +76,7 @@
             this.RecursiveSearchBox.Location = new System.Drawing.Point(8, 99);
             this.RecursiveSearchBox.Name = "RecursiveSearchBox";
             this.RecursiveSearchBox.Size = new System.Drawing.Size(141, 20);
-            this.RecursiveSearchBox.TabIndex = 6;
-            this.RecursiveSearchBox.TabStop = false;
+            this.RecursiveSearchBox.TabIndex = 8;
             this.RecursiveSearchBox.Text = "Recursive &search?";
             this.RecursiveSearchBox.UseVisualStyleBackColor = true;
             this.RecursiveSearchBox.Visible = false;
@@ -89,7 +89,7 @@
             this.FindReplaceFormTitle.Location = new System.Drawing.Point(83, 9);
             this.FindReplaceFormTitle.Name = "FindReplaceFormTitle";
             this.FindReplaceFormTitle.Size = new System.Drawing.Size(201, 22);
-            this.FindReplaceFormTitle.TabIndex = 2;
+            this.FindReplaceFormTitle.TabIndex = 13;
             this.FindReplaceFormTitle.Text = "Find/replace in JSON";
             // 
             // FindButton
@@ -98,7 +98,7 @@
             this.FindButton.Location = new System.Drawing.Point(87, 359);
             this.FindButton.Name = "FindButton";
             this.FindButton.Size = new System.Drawing.Size(86, 23);
-            this.FindButton.TabIndex = 8;
+            this.FindButton.TabIndex = 11;
             this.FindButton.Text = "Find all";
             this.FindButton.UseVisualStyleBackColor = true;
             this.FindButton.Click += new System.EventHandler(this.FindButton_Click);
@@ -110,7 +110,7 @@
             this.ReplaceButton.Location = new System.Drawing.Point(187, 359);
             this.ReplaceButton.Name = "ReplaceButton";
             this.ReplaceButton.Size = new System.Drawing.Size(97, 23);
-            this.ReplaceButton.TabIndex = 9;
+            this.ReplaceButton.TabIndex = 12;
             this.ReplaceButton.Text = "Replace all";
             this.ReplaceButton.UseVisualStyleBackColor = true;
             this.ReplaceButton.Click += new System.EventHandler(this.ReplaceButton_Click);
@@ -132,7 +132,7 @@
             this.FindTextBoxLabel.Location = new System.Drawing.Point(242, 55);
             this.FindTextBoxLabel.Name = "FindTextBoxLabel";
             this.FindTextBoxLabel.Size = new System.Drawing.Size(42, 16);
-            this.FindTextBoxLabel.TabIndex = 9;
+            this.FindTextBoxLabel.TabIndex = 14;
             this.FindTextBoxLabel.Text = "Find...";
             // 
             // RegexBox
@@ -143,8 +143,7 @@
             this.RegexBox.Location = new System.Drawing.Point(8, 59);
             this.RegexBox.Name = "RegexBox";
             this.RegexBox.Size = new System.Drawing.Size(182, 20);
-            this.RegexBox.TabIndex = 4;
-            this.RegexBox.TabStop = false;
+            this.RegexBox.TabIndex = 6;
             this.RegexBox.Text = "Use re&gular expressions?";
             this.RegexBox.UseVisualStyleBackColor = true;
             this.RegexBox.Visible = false;
@@ -157,8 +156,7 @@
             this.MathBox.Location = new System.Drawing.Point(201, 59);
             this.MathBox.Name = "MathBox";
             this.MathBox.Size = new System.Drawing.Size(127, 20);
-            this.MathBox.TabIndex = 5;
-            this.MathBox.TabStop = false;
+            this.MathBox.TabIndex = 7;
             this.MathBox.Text = "&Math expression";
             this.MathBox.UseVisualStyleBackColor = true;
             this.MathBox.Visible = false;
@@ -170,7 +168,7 @@
             this.ReplaceTextBox.Location = new System.Drawing.Point(31, 91);
             this.ReplaceTextBox.Name = "ReplaceTextBox";
             this.ReplaceTextBox.Size = new System.Drawing.Size(196, 22);
-            this.ReplaceTextBox.TabIndex = 1;
+            this.ReplaceTextBox.TabIndex = 2;
             this.ReplaceTextBox.KeyDown += new System.Windows.Forms.KeyEventHandler(this.FindReplaceForm_KeyDown);
             this.ReplaceTextBox.KeyPress += new System.Windows.Forms.KeyPressEventHandler(this.TextBox_KeyPress);
             this.ReplaceTextBox.KeyUp += new System.Windows.Forms.KeyEventHandler(this.FindReplaceForm_KeyUp);
@@ -181,7 +179,7 @@
             this.ReplaceTextBoxLabel.Location = new System.Drawing.Point(242, 94);
             this.ReplaceTextBoxLabel.Name = "ReplaceTextBoxLabel";
             this.ReplaceTextBoxLabel.Size = new System.Drawing.Size(93, 16);
-            this.ReplaceTextBoxLabel.TabIndex = 13;
+            this.ReplaceTextBoxLabel.TabIndex = 15;
             this.ReplaceTextBoxLabel.Text = "Replace with...";
             // 
             // KeysValsBothBoxLabel
@@ -190,7 +188,7 @@
             this.KeysValsBothBoxLabel.Location = new System.Drawing.Point(152, 24);
             this.KeysValsBothBoxLabel.Name = "KeysValsBothBoxLabel";
             this.KeysValsBothBoxLabel.Size = new System.Drawing.Size(160, 16);
-            this.KeysValsBothBoxLabel.TabIndex = 14;
+            this.KeysValsBothBoxLabel.TabIndex = 18;
             this.KeysValsBothBoxLabel.Text = "Search in keys or values?";
             this.KeysValsBothBoxLabel.Visible = false;
             // 
@@ -201,7 +199,7 @@
             this.RootTextBox.Location = new System.Drawing.Point(31, 130);
             this.RootTextBox.Name = "RootTextBox";
             this.RootTextBox.Size = new System.Drawing.Size(196, 22);
-            this.RootTextBox.TabIndex = 2;
+            this.RootTextBox.TabIndex = 3;
             this.RootTextBox.KeyDown += new System.Windows.Forms.KeyEventHandler(this.FindReplaceForm_KeyDown);
             this.RootTextBox.KeyPress += new System.Windows.Forms.KeyPressEventHandler(this.TextBox_KeyPress);
             this.RootTextBox.KeyUp += new System.Windows.Forms.KeyEventHandler(this.FindReplaceForm_KeyUp);
@@ -227,7 +225,7 @@
             this.AdvancedGroupBox.Location = new System.Drawing.Point(24, 180);
             this.AdvancedGroupBox.Name = "AdvancedGroupBox";
             this.AdvancedGroupBox.Size = new System.Drawing.Size(337, 163);
-            this.AdvancedGroupBox.TabIndex = 17;
+            this.AdvancedGroupBox.TabIndex = 4;
             this.AdvancedGroupBox.TabStop = false;
             // 
             // MatchExactlyBox
@@ -237,11 +235,11 @@
             this.MatchExactlyBox.Location = new System.Drawing.Point(96, 130);
             this.MatchExactlyBox.Name = "MatchExactlyBox";
             this.MatchExactlyBox.Size = new System.Drawing.Size(117, 20);
-            this.MatchExactlyBox.TabIndex = 16;
-            this.MatchExactlyBox.TabStop = false;
+            this.MatchExactlyBox.TabIndex = 10;
             this.MatchExactlyBox.Text = "Match &exactly?";
             this.MatchExactlyBox.UseVisualStyleBackColor = true;
             this.MatchExactlyBox.CheckedChanged += new System.EventHandler(this.MatchExactlyBox_CheckedChanged);
+            this.MatchExactlyBox.KeyUp += new System.Windows.Forms.KeyEventHandler(this.FindReplaceForm_KeyUp);
             // 
             // IgnoreCaseCheckBox
             // 
@@ -249,39 +247,41 @@
             this.IgnoreCaseCheckBox.Location = new System.Drawing.Point(201, 99);
             this.IgnoreCaseCheckBox.Name = "IgnoreCaseCheckBox";
             this.IgnoreCaseCheckBox.Size = new System.Drawing.Size(107, 20);
-            this.IgnoreCaseCheckBox.TabIndex = 15;
-            this.IgnoreCaseCheckBox.TabStop = false;
+            this.IgnoreCaseCheckBox.TabIndex = 9;
             this.IgnoreCaseCheckBox.Text = "Ignore &case?";
             this.IgnoreCaseCheckBox.UseVisualStyleBackColor = true;
-            // 
-            // AdvancedGroupBoxLabel
-            // 
-            this.AdvancedGroupBoxLabel.AutoSize = true;
-            this.AdvancedGroupBoxLabel.Location = new System.Drawing.Point(28, 170);
-            this.AdvancedGroupBoxLabel.Name = "AdvancedGroupBoxLabel";
-            this.AdvancedGroupBoxLabel.Size = new System.Drawing.Size(151, 16);
-            this.AdvancedGroupBoxLabel.TabIndex = 15;
-            this.AdvancedGroupBoxLabel.Text = "Show advanced options";
-            this.AdvancedGroupBoxLabel.Click += new System.EventHandler(this.AdvancedGroupBoxLabel_Click);
+            this.IgnoreCaseCheckBox.KeyUp += new System.Windows.Forms.KeyEventHandler(this.FindReplaceForm_KeyUp);
             // 
             // SwapFindReplaceButton
             // 
             this.SwapFindReplaceButton.Location = new System.Drawing.Point(303, 60);
             this.SwapFindReplaceButton.Name = "SwapFindReplaceButton";
             this.SwapFindReplaceButton.Size = new System.Drawing.Size(49, 28);
-            this.SwapFindReplaceButton.TabIndex = 18;
+            this.SwapFindReplaceButton.TabIndex = 1;
             this.SwapFindReplaceButton.Text = "Swap";
             this.SwapFindReplaceButton.UseVisualStyleBackColor = true;
             this.SwapFindReplaceButton.Click += new System.EventHandler(this.SwapFindReplaceButton_Click);
             this.SwapFindReplaceButton.KeyUp += new System.Windows.Forms.KeyEventHandler(this.FindReplaceForm_KeyUp);
+            // 
+            // ShowAdvancedOptionsCheckBox
+            // 
+            this.ShowAdvancedOptionsCheckBox.AutoSize = true;
+            this.ShowAdvancedOptionsCheckBox.Location = new System.Drawing.Point(32, 170);
+            this.ShowAdvancedOptionsCheckBox.Name = "ShowAdvancedOptionsCheckBox";
+            this.ShowAdvancedOptionsCheckBox.Size = new System.Drawing.Size(173, 20);
+            this.ShowAdvancedOptionsCheckBox.TabIndex = 4;
+            this.ShowAdvancedOptionsCheckBox.Text = "Show advanced options";
+            this.ShowAdvancedOptionsCheckBox.UseVisualStyleBackColor = true;
+            this.ShowAdvancedOptionsCheckBox.CheckedChanged += new System.EventHandler(this.ShowAdvancedOptionsCheckBox_CheckedChanged);
+            this.ShowAdvancedOptionsCheckBox.KeyUp += new System.Windows.Forms.KeyEventHandler(this.FindReplaceForm_KeyUp);
             // 
             // FindReplaceForm
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(8F, 16F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
             this.ClientSize = new System.Drawing.Size(373, 394);
+            this.Controls.Add(this.ShowAdvancedOptionsCheckBox);
             this.Controls.Add(this.SwapFindReplaceButton);
-            this.Controls.Add(this.AdvancedGroupBoxLabel);
             this.Controls.Add(this.AdvancedGroupBox);
             this.Controls.Add(this.RootTextBoxLabel);
             this.Controls.Add(this.RootTextBox);
@@ -321,9 +321,9 @@
         private System.Windows.Forms.TextBox RootTextBox;
         private System.Windows.Forms.Label RootTextBoxLabel;
         private System.Windows.Forms.GroupBox AdvancedGroupBox;
-        private System.Windows.Forms.Label AdvancedGroupBoxLabel;
         private System.Windows.Forms.Button SwapFindReplaceButton;
         private System.Windows.Forms.CheckBox IgnoreCaseCheckBox;
         private System.Windows.Forms.CheckBox MatchExactlyBox;
+        private System.Windows.Forms.CheckBox ShowAdvancedOptionsCheckBox;
     }
 }