+++ JsonToolsNppPlugin/Forms/FindReplaceForm.cs
@@ -56,6 +56,18 @@ namespace JSON_Tools.Forms
                 e.SuppressKeyPress = true;
         }
 
+        /// <summary>
+        /// suppress the default response to the Tab key
+        /// </summary>
+        /// <param name="keyData"></param>
+        /// <returns></returns>
+        protected override bool ProcessDialogKey(Keys keyData)
+        {
+            if (keyData.HasFlag(Keys.Tab)) // this covers Tab with or without modifiers
+                return true;
+            return base.ProcessDialogKey(keyData);
+        }
+
         private void FindReplaceForm_KeyUp(object sender, KeyEventArgs e)
         {
             SortForm.GenericKeyUpHandler(this, sender, e);
@@ -76,14 +88,14 @@ namespace JSON_Tools.Forms
                 e.Handled = true;
         }
 
-        private void AdvancedGroupBoxLabel_Click(object sender, EventArgs e)
+        private void ShowAdvancedOptionsCheckBox_CheckedChanged(object sender, EventArgs e)
         {
             // show the advanced controls and expand the box
             if (!RegexBox.Visible)
             {
                 if (Height < EXTENDED_HEIGHT)
                     Height = EXTENDED_HEIGHT;
-                AdvancedGroupBoxLabel.Text = "Hide advanced options";
+                ShowAdvancedOptionsCheckBox.Text = "Hide advanced options";
                 AdvancedGroupBox.Height = ADVANCED_CONTROLS_EXTENDED_HEIGHT;
                 foreach (Control control in AdvancedGroupBox.Controls)
                 {
@@ -95,7 +107,7 @@ namespace JSON_Tools.Forms
             {
                 if (Height == EXTENDED_HEIGHT)
                     Height = COLLAPSED_HEIGHT;
-                AdvancedGroupBoxLabel.Text = "Show advanced options";
+                ShowAdvancedOptionsCheckBox.Text = "Show advanced options";
                 AdvancedGroupBox.Height = ADVANCED_CONTROLS_COLLAPSED_HEIGHT;
                 foreach (Control control in AdvancedGroupBox.Controls)
                 {
@@ -233,11 +245,15 @@ namespace JSON_Tools.Forms
             // for math find/replace we want to filter before performing the operation
             // for regex find/replace the s_sub function naturally filters (because it's only replacing the target substring)
             // thus, we only want to use the findQuery to filter when doing math find/replace
-            treeViewer.QueryBox.Text = (MathBox.Checked)
-                ? findQuery + " = " + replaceQuery
-                : $"{root}[is_str(@)] = {replaceQuery}";
-            //treeViewer.QueryBox.Text = findQuery + " = " + replaceQuery;
+            string findPart = MathBox.Checked
+                ? $"var x = {findQuery}"
+                : $"var x = {root}[is_str(@)]";
+            treeViewer.QueryBox.Text = $"{findPart};\r\nx = {replaceQuery};\r\nx";
+            // use variable assignment, then mutate the variable, then show the variable.
+            // this is nice because (1) it introduces variable assignment,
+            // and (2) it displays exactly the values that were mutated rather than forcing the user to find them
             treeViewer.SubmitQueryButton.PerformClick();
+            treeViewer.Tree.Nodes[0].Expand();
             KeysValsBothBox.SelectedIndex = keysvals_index;
         }